place: move initial placement in separate file
