+++ vpr/src/place/initial_placement.h
@@ -1,0 +1,9 @@
+#ifndef VPR_INITIAL_PLACEMENT_H
+#define VPR_INITIAL_PLACEMENT_H
+
+#include "vpr_types.h"
+
+void initial_placement(enum e_pad_loc_type pad_loc_type,
+                       const char* pad_loc_file);
+
+#endif
